+++ plugins/modules/zhmc_hba.py
@@ -638,7 +638,7 @@ def main():
     # by showing the traceback.
 
     LOGGER.debug(
-        "Module exit (success): changed: %r, cpc: %r", changed, result)
+        "Module exit (success): changed: %r, hba: %r", changed, result)
     module.exit_json(changed=changed, hba=result)