+++ plugins/modules/zhmc_storage_group.py
@@ -1136,7 +1136,8 @@ def main():
     # by showing the traceback.
 
     LOGGER.debug(
-        "Module exit (success): changed: %r, cpc: %r", changed, result)
+        "Module exit (success): changed: %r, storage_group: %r",
+        changed, result)
     module.exit_json(changed=changed, storage_group=result)